A key struct for moving string escaping out of loops.
The cases where high performance JSON writing is needed the most
is likely to include looping over objects of the same type and
repeatedly writing them to the JSON stream. Most keys will be
used over and over again. To take advantage of this we can
introduce a new key struct that can be used where raw strings
are now used in the property pairs. It moves the string escaping
code out of expensive loops.