Control the inlining of require_bytes more carefully.
Since require_bytes needs to be very fast, it makes sense to force
inlining of it, but keep it very short, so that it does not bloat the
methods calling it. In this case it means that we need to keep
grow_buffer from being inlined with it, which is ok, since it
grow_buffer is called, the majority of the time will be taken by the
realloc call.