cos: Order initializer lists
C++ will initialize members in the order they are declared in the
class. The order in the initialization list may be different, but it
doesn't affect construction or destruction order. Thus it is nice if
they are in the same order, as weird bugs might occur otherwise.

There is a nice warning that we currently have disabled that catches
this, it would be nice if we could enable it.

Change-Id: I44fc9d0f7208fa5ce888009101a04afa0712fb6f