Add some utility methods to the encoding_context
These are not strictly necessary, since you can do the same
thing using the public method alread, but they do save a lot
of code at the call site, when the caller does not have to
reserve and advance to write a single little byte to the
context. Replacing the last written byte is also possible
by reserving 0 bytes and stepping back one byte from the
returned pointer, but that is not very obvious; so better to
add this functionality to this class instead.