Remove templates from string escaper
This is quite a bit of code that it would be best to not repeat for
different output types, that is not have the compiler specialize
several times. Just limit the code to working with pointers to a
byte buffer.

As an bonus, this actually sped up the string benchmarks a bit.
Not sure why, but maybe the compiler didn't apply as agressive
optimizations before.