Fix performance regression on macOS (or Clang)
For some reason, putting the string skipping functions in
the .cpp file compiles as we expect and get much better performance
out of the SSE version. Putting this code in the header file
actually makes the performance worse than the non-SSE version,
but only on macOS (or perhaps on Clang, not tried on Linux).