Fix performance problem due to uninitialized variable
This bug has been here for some time, but hasn't been caught by
tests or in production, because if the value of this unsigned
variable is less than the inline size, then we use parts of the
inline buffer before switching to a heap allocated buffer. If
we instead get a random value that is way larger than the
inline size, then we immediately switch over to the heap
allocated vector. I would guess that the latter case is the most
common one.