Remove problematic unit test
This unit test has proven a bit tricky to implement and keep
problem free. When running with address sanitizer, we need to
tell AS to not worry about the huge allocation. Valgrind will
assume a mistake in the code when seeing the huge allocation.
Finally, on Windows we seem to end up in a slightly problematic
state after the failed allocation, which I do not believe is
our fault. So let's remove this test. It seems fairly unlikely
that we will accidentally remove the bad_alloc if-statemen.