Git rid of false positive from Clang static analyzer
The code is correct, since the nullptr begin will be equal to the nullptr
end, and the parsing code (further down the call chain) will exit instead
of dereferencing the pointer. This is a bit too much for the Clang static
analyzer to figure out though (it's a pretty deep call chain for the false
positive it finds), so I'm changing the implementation a bit to make it
easier to analyze. The functionality is identical.