Disable Clang static analyzer for the double-conversion library
This is provided by a 3rd party vendor (Google) and not under our control.
Not sure if the problem found in this library by the analyzer is real or
not, but this code is used in quite high profile codebases, such as
Chrome, so it is probably pretty battle tested.