Fix parsing JSON data with null bytes
The same problematic _mm_cmpistri intrinsic was used when decoding as
well. This could cause the decoder to not look for quotes or backslashes
in chunks with null bytes in them.

Using the wrong intrinsic in the skip_any_whitespace case doesn't
actally cause any issues since the end-of-string is treated as "not
whitespace". For clarity, it was changed to _mm_cmpestri there too.